Backend/momment/requirements.txt
appnope==0.1.3
asgiref==3.5.2
astroid==2.12.13
asttokens==2.0.5
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.4
cryptography==38.0.3
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
dj-rest-auth==2.2.5
Django==3.2.13
django-allauth==0.50.0
django-appconf==1.0.5
django-bootstrap-v5==1.0.11
django-bootstrap4==22.3
django-cors-headers==3.13.0
django-extensions==3.2.1
django-imagekit==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.0.4
executing==1.1.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
ipython==8.5.0
isort==5.11.4
jedi==0.18.1
lazy-object-proxy==1.8.0
matplotlib-inline==0.1.6
mccabe==0.7.0
oauthlib==3.2.2
outcome==1.2.0
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==9.2.0
platformdirs==2.6.2
prompt-toolkit==3.0.31
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
pylint==2.15.9
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
selenium==4.6.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.5.1
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
traitlets==5.4.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
webdriver-manager==3.8.5
wrapt==1.14.1
wsproto==1.2.0
